requirements.txt
Flask>=2.2
Flask-Login>=0.6
Flask-WTF>=1.1
Flask-Migrate>=4.0
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
requests>=2.28
werkzeug>=2.2
